+++ scripts/helperFunctions.ts
@@ -1,13 +1,12 @@
 import { Signer } from "@ethersproject/abstract-signer";
 import { Contract } from "@ethersproject/contracts";
 import { Network } from "hardhat/types";
+import { alchemica, maticDiamondAddress } from "../constants";
 import {
   AlchemicaToken,
   DiamondLoupeFacet,
   OwnershipFacet,
 } from "../typechain";
-
-export const gasPrice = 100000000000;
 
 export async function impersonate(
   address: string,
@@ -61,24 +60,6 @@ export function getSelector(func: string, ethers: any) {
   const abiInterface = new ethers.utils.Interface([func]);
   return abiInterface.getSighash(ethers.utils.Fragment.from(func));
 }
-
-export const kovanDiamondAddress = "0xa37D0c085121B6b7190A34514Ca28fC15Bb4dc22";
-export const maticDiamondAddress = "0x1D0360BaC7299C86Ec8E99d0c1C9A95FEfaF2a11";
-export const maticAavegotchiDiamondAddress =
-  "0x86935F11C86623deC8a25696E1C19a8659CbF95d";
-export const aavegotchiDAOAddress =
-  "0xb208f8BB431f580CC4b216826AFfB128cd1431aB";
-export const pixelcraftAddress = "0xD4151c984e6CF33E04FFAAF06c3374B2926Ecc64";
-
-export const alchemica = [
-  "0x403E967b044d4Be25170310157cB1A4Bf10bdD0f", //fud
-  "0x44A6e0BE76e1D9620A7F76588e4509fE4fa8E8C8", //fomo
-  "0x6a3E7C3c6EF65Ee26975b12293cA1AAD7e1dAeD2", //alpha
-  "0x42E5E06EF5b90Fe15F853F59299Fc96259209c5C", //kek
-];
-
-export const ecosystemVesting = "0x7e07313B4FF259743C0c84eA3d5e741D2b0d07c3";
-export const gameplayVesting = "0x3fB6C2A83d2FfFe94e0b912b612fB100047cc176";
 
 export async function diamondOwner(address: string, ethers: any) {
   return await (await ethers.getContractAt("OwnershipFacet", address)).owner();
