+++ scripts/tile/addTileTypes.ts
@@ -1,41 +1,16 @@
-import { BigNumber, Signer } from "ethers";
+import { Signer } from "ethers";
 import { ethers } from "hardhat";
 import { tileTypes } from "../../data/tiles/tileTypes";
 import { TileFacet, OwnershipFacet } from "../../typechain";
-import { TileTypeInput, TileTypeOutput } from "../../types";
+
 import {
   aavegotchiDAOAddress,
   gasPrice,
   maticAavegotchiDiamondAddress,
   maticDiamondAddress,
   pixelcraftAddress,
-} from "../helperFunctions";
-
-function outputTile(tile: TileTypeInput): TileTypeOutput {
-  if (tile.width > 64) throw new Error("Width too much");
-  if (tile.height > 64) throw new Error("Height too much");
-
-  const alchemica = tile.alchemicaCost.map((val) =>
-    ethers.utils.parseEther(val.toString())
-  );
-
-  let output: TileTypeOutput = {
-    width: tile.width,
-    height: tile.height,
-    deprecated: tile.deprecated,
-    tileType: tile.tileType,
-    craftTime: tile.craftTime,
-    alchemicaCost: [
-      BigNumber.from(alchemica[0]),
-      BigNumber.from(alchemica[1]),
-      BigNumber.from(alchemica[2]),
-      BigNumber.from(alchemica[3]),
-    ],
-    name: tile.name,
-  };
-
-  return output;
-}
+} from "../../constants";
+import { outputTile } from "./helperFunctions";
 
 export async function setAddresses() {
   const accounts: Signer[] = await ethers.getSigners();
@@ -59,21 +34,21 @@ export async function setAddresses() {
     deployer
   )) as TileFacet;
 
-  // // add real data
-  // const goldenTiles = tileTypes.map((val) => outputTile(val));
+  // add real data
+  const goldenTiles = tileTypes.map((val) => outputTile(val, ethers));
 
-  // await tileFacet.addTileTypes(goldenTiles, {
-  //   gasPrice: gasPrice,
-  // });
+  await tileFacet.addTileTypes(goldenTiles, {
+    gasPrice: gasPrice,
+  });
 
-  // await tileFacet.setAddresses(
-  //   maticAavegotchiDiamondAddress,
-  //   maticDiamondAddress,
-  //   ethers.constants.AddressZero,
-  //   pixelcraftAddress,
-  //   aavegotchiDAOAddress,
-  //   { gasPrice: gasPrice }
-  // );
+  await tileFacet.setAddresses(
+    maticAavegotchiDiamondAddress,
+    maticDiamondAddress,
+    ethers.constants.AddressZero,
+    pixelcraftAddress,
+    aavegotchiDAOAddress,
+    { gasPrice: gasPrice }
+  );
 
   const tiles = await tileFacet.getTileTypes([]);
   console.log("tiles:", tiles);
